Teacher course details

+++ app/Http/Controllers/TeacherController.php
@@ -2,10 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Course;
 use App\Http\Requests\TeacherProfileRequest;
 use App\Http\Requests\ProPicRequest;
+use App\Student;
 use Illuminate\Http\Request;
 use App\Teacher;
+use App\TeacherCourse;
 
 class TeacherController extends Controller
 {
@@ -14,8 +17,13 @@ class TeacherController extends Controller
     {
         $teacher = Teacher::where('username',$request->session()->get('username'))
                     ->first();
+        $teacherCourse = TeacherCourse::where('teacher_id',$teacher->teacher_id)
+                                       ->where('status','In progress')
+                                       ->get();
 
-        return view('teacher.index',compact('teacher'));
+        $course = Course::all();
+
+        return view('teacher.index',compact('teacher','course','teacherCourse'));
     }
 
     public function profile(Request $request)
@@ -51,25 +59,42 @@ class TeacherController extends Controller
     {
         $teacher = Teacher::where('username',$request->session()->get('username'))
                     ->first();
-        if($teacher->profile_pic) {
-            unlink($teacher->profile_pic);
+
+        if($request->profile_pic)
+        {
+            if($teacher->profile_pic) {
+                unlink($teacher->profile_pic);
+            }
+            $extension = $request->profile_picture->getClientOriginalExtension();
+            $fileName = date('U').'.'.$extension;
+            $destination = "images/teacher/";
+            $request->profile_picture->move($destination, $fileName);
+            $teacher->profile_pic = $destination.$fileName;
+
         }
-        $extension = $request->profile_picture->getClientOriginalExtension();
-        $fileName = date('U').'.'.$extension;
-        $destination = "images/teacher/";
-        $request->profile_picture->move($destination, $fileName);
-        $teacher->profile_pic = $destination.$fileName;
 
         $teacher->name = $request->name;
         $teacher->phone = $request->phone;
         $teacher->email = $request->email;
-        $teacher->birthdate = $request->birthdate;
-        $teacher->status = $request->status;
         $teacher->address = $request->address;
 
         $teacher->save();
         $request->session()->flash('msg','Profile Updated Sucessfully!');
         return view('teacher.profile',compact('teacher'));
+
     }
+
+    public function viewStudent(Request $request)
+    {
+        $teacher = Teacher::where('username',$request->session()->get('username'))
+                    ->first();
+        $teacherCourse = TeacherCourse::where('teacher_id',$teacher->teacher_id)
+                                       ->get();
+
+        $course = Course::all();
+        $student = Student::all();
+        return view('teacher.viewStudent',compact('teacher','course','teacherCourse'));
+    }
+
 
 }
